requirements.txt
numpy
numpy-stl
pillow